backend/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
flask-restx==1.0.3
boto3==1.35.48
botocore==1.31.48
python-dotenv==1.0.0
werkzeug==2.0.1
SQLAlchemy==1.4.23
pydantic==2.5.2
pydantic-settings==2.1.0



